refactor: 패키지 구조 수정 formatter 기능의 객체들을 domain 영역이 아닌 view 영역으로 이동
