+++ src/main/java/racingcar/RacingCarContext.java
@@ -2,10 +2,10 @@ package racingcar;
 
 import racingcar.controller.RacingController;
 import racingcar.model.car.CarFactory;
-import racingcar.model.car.formatter.CarPositionFormatter;
-import racingcar.model.car.formatter.CarsPositionFormatter;
+import racingcar.view.formatter.car.CarPositionFormatter;
+import racingcar.view.formatter.car.CarsPositionFormatter;
 import racingcar.model.car.strategy.MovingStrategy;
-import racingcar.model.track.formatter.TrackStateFormatter;
+import racingcar.view.formatter.track.TrackStateFormatter;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
